requirements.txt
Pillow>=9.0.1
PySimpleGUI>=4.57.0
requests>=2.27.1
selenium >= 4.1.3
